requirements.txt
telethon==1.36.0
aiogram==3.15.0
aiosqlite==0.20.0
google-generativeai==0.8.4
pillow==11.1.0
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
aiohttp>=3.9.0,<3.11
